repo sync: Don't complain about deleted remote branches
Right now, if you run `git pull` or `git fetch` with `--prune`
before calling `repo sync`,
it will complain that the remote tracking branch is missing.
It's only an error, not a warning, but it can still be confusing.

This makes `repo sync` check if the remote tracking branch exists
before trying to delete it,
making it a no-op if the branch has already been deleted.

+++ repo_sync.go
@@ -398,12 +398,15 @@ func (cmd *repoSyncCmd) deleteMergedBranches(
 			return fmt.Errorf("delete branch %v: %w", branch, err)
 		}
 
-		// Also delete the remote tracking branch for this branch.
+		// Also delete the remote tracking branch for this branch
+		// if it still exists.
 		remoteBranch := remote + "/" + branch
-		if err := repo.DeleteBranch(ctx, remoteBranch, git.BranchDeleteOptions{
-			Remote: true,
-		}); err != nil {
-			log.Warn("Unable to delete remote tracking branch", "branch", remoteBranch, "error", err)
+		if _, err := repo.PeelToCommit(ctx, remoteBranch); err == nil {
+			if err := repo.DeleteBranch(ctx, remoteBranch, git.BranchDeleteOptions{
+				Remote: true,
+			}); err != nil {
+				log.Warn("Unable to delete remote tracking branch", "branch", remoteBranch, "error", err)
+			}
 		}
 	}
